Reorganize warp drive patches
