+++ NewHorizons/Tools/Patches.cs
@@ -50,13 +50,8 @@ namespace NewHorizons.Tools
             Main.Instance.ModHelper.HarmonyHelper.AddPrefix<ProbeLauncher>("UpdateOrbitalLaunchValues", typeof(Patches), nameof(Patches.OnProbeLauncherUpdateOrbitalLaunchValues));
             Main.Instance.ModHelper.HarmonyHelper.AddPrefix<SurveyorProbe>("IsLaunched", typeof(Patches), nameof(Patches.OnSurveyorProbeIsLaunched));
 
-            Main.Instance.ModHelper.HarmonyHelper.AddPrefix<ShipLogController>("Update", typeof(Patches), nameof(Patches.OnShipLogControllerUpdate));
-
-            Main.Instance.ModHelper.HarmonyHelper.AddPrefix<ShipCockpitController>("Update", typeof(Patches), nameof(Patches.OnShipCockpitControllerUpdate));
-
             // Postfixes
             Main.Instance.ModHelper.HarmonyHelper.AddPostfix<MapController>("Awake", typeof(Patches), nameof(Patches.OnMapControllerAwake));
-            Main.Instance.ModHelper.HarmonyHelper.AddPostfix<ShipLogMapMode>("EnterMode", typeof(Patches), nameof(Patches.OnShipLogMapModeEnterMode));
         }
 
         public static bool GetHUDDisplayName(ReferenceFrame __instance, ref string __result)
@@ -340,38 +335,6 @@ namespace NewHorizons.Tools
             return (Locator.GetPlayerRulesetDetector()?.GetPlanetoidRuleset()?.GetGravityVolume() != null);
         }
 
-        public static bool OnShipLogControllerUpdate(ShipLogController __instance)
-        {
-            if (__instance._exiting
-                || OWInput.GetInputMode() != InputMode.ShipComputer
-                || __instance._currentMode.AllowCancelInput() && OWInput.IsNewlyPressed(InputLibrary.cancel, InputMode.All)
-                || ShipLogBuilder.ShipLogStarChartMode == null)
-                return true;
-
-            __instance._exitPrompt.SetVisibility(__instance._currentMode.AllowCancelInput());
-            __instance._currentMode.UpdateMode();
-            if (__instance._currentMode.AllowModeSwap() && OWInput.IsNewlyPressed(InputLibrary.swapShipLogMode, InputMode.All))
-            {
-                ShipLogMode currentMode = __instance._currentMode;
-                string focusedEntryID = currentMode.GetFocusedEntryID();
-                if (!focusedEntryID.Equals("")) return true;
-                bool flag = currentMode.Equals(__instance._mapMode);
-                __instance._currentMode = (flag ? __instance._detectiveMode : __instance._mapMode);
-
-                if (currentMode.Equals(__instance._mapMode))
-                    __instance._currentMode = ShipLogBuilder.ShipLogStarChartMode;
-                else if (currentMode.Equals(ShipLogBuilder.ShipLogStarChartMode))
-                    __instance._currentMode = __instance._detectiveMode;
-                else
-                    __instance._currentMode = __instance._mapMode;
-
-                currentMode.ExitMode();
-                __instance._currentMode.EnterMode(focusedEntryID, null);
-                __instance._oneShotSource.PlayOneShot(flag ? global::AudioType.ShipLogEnterDetectiveMode : global::AudioType.ShipLogEnterMapMode, 1f);
-            }
-            return false;
-        }
-
         public static bool OnSurveyorProbeIsLaunched(SurveyorProbe __instance, ref bool __result)
         {
             try
@@ -383,28 +346,6 @@ namespace NewHorizons.Tools
                 __result = true;
             }
             return false;
-        }
-
-        public static bool OnShipCockpitControllerUpdate(ShipCockpitController __instance)
-        {
-            if(__instance._playerAtFlightConsole && OWInput.IsNewlyPressed(InputLibrary.autopilot, InputMode.ShipCockpit))
-            {
-                var targetSystem = ShipLogBuilder.ShipLogStarChartMode.GetTargetStarSystem();
-                if (targetSystem != null)
-                {
-                    Main.Instance.ChangeCurrentStarSystem(targetSystem, true);
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        public static void OnShipLogMapModeEnterMode(ShipLogMapMode __instance)
-        {
-            var newPrompt = "Interstellar Mode";
-            __instance._detectiveModePrompt.SetText(newPrompt);
-            var text = GameObject.Find("Ship_Body/Module_Cabin/Systems_Cabin/ShipLogPivot/ShipLog/ShipLogPivot/ShipLogCanvas/ScreenPromptListScaleRoot/ScreenPromptList_UpperRight/ScreenPrompt/Text").GetComponent<UnityEngine.UI.Text>();
-            text.text = newPrompt;
         }
     }
 }
